# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Contactless cashless event payment system API (Colombia market)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{eventId}/access-zones:
    get:
      tags: [AccessZones]
      operationId: listAccessZones
      summary: List access zones for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of access zones.
          content:
            application/json:
              schema:
                type: object
                required: [zones]
                properties:
                  zones:
                    type: array
                    items:
                      $ref: '#/components/schemas/AccessZone'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [AccessZones]
      operationId: createAccessZone
      summary: Create an access zone for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccessZoneBody'
      responses:
        '201':
          description: Access zone created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessZone'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /events/{eventId}/access-zones/{zoneId}:
    patch:
      tags: [AccessZones]
      operationId: updateAccessZone
      summary: Update an access zone
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
        - name: zoneId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccessZoneBody'
      responses:
        '200':
          description: Access zone updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessZone'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [AccessZones]
      operationId: deleteAccessZone
      summary: Delete an access zone
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
        - name: zoneId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Access zone deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  # ── Push Tokens ───────────────────────────────────────────────────────────────
...
components:
...
  responses:
    BadRequest:
      description: Invalid request body or parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Not authenticated.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Resource already exists.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
...
    AccessZone:
      type: object
      required: [id, eventId, name, rank, createdAt]
      properties:
        id:
          type: string
        eventId:
          type: string
        name:
          type: string
        description:
          type: [string, 'null']
        colorHex:
          type: [string, 'null']
        rank:
          type: integer
        upgradePrice:
          type: [integer, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateAccessZoneBody:
      type: object
      required: [name, rank]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        colorHex:
          type: string
        rank:
          type: integer
          minimum: 0
        upgradePrice:
          type: [integer, 'null']
    UpdateAccessZoneBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        colorHex:
          type: string
        rank:
          type: integer
          minimum: 0
        upgradePrice:
          type: [integer, 'null']